# Optimize type canonicalization